fix vec4 compiler warnings

+++ libs/openFrameworks/math/ofVec4f.h
@@ -39,11 +39,11 @@ public:
 		return (const float *)&x;
 	}
 	
-	float& operator[]( int n ){
+	float& operator[]( size_t n ){
 		return getPtr()[n];
 	}
 	
-	float operator[]( int n ) const {
+	float operator[]( size_t n ) const {
 		return getPtr()[n];
 	}
